feat: Add reciever package for consuming messages
The code changes introduce a new package called "reciever" for consuming messages. This package is used in the main.go file to start consuming messages from RabbitMQ. By separating the consuming logic into its own package, the code becomes more modular and easier to maintain. This enhancement improves the overall architecture of the backend application.
